Operators

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -1,8 +1,12 @@
 package lang.print.gaps.task5;
 
 public class NumberDivider {
-    void divide(int toBeDivided) {
-        double result = toBeDivided/5;
+    void divide(float toBeDivided) {
+        float result = toBeDivided/5;
         System.out.println(result);
     }
 }
+
+
+
+
